requirements.txt
streamlit
pandas
plotly
statsmodels
google-cloud-firestore